Move get_all_cluster_yaml_files function
- From helm_upgrade_decission.py to file_acquisition.py
- Because it is not only used by helm upgrade cmd,
  it makes more sense to be in this utility file instead.
